etl_pipeline/slack_bot/requirements.txt
sqlalchemy
psycopg2-binary
requests
python-dotenv